Fixes related to #1
