+++ asabiris/output/smtp/service.py
@@ -15,7 +15,7 @@ L = logging.getLogger(__name__)
 asab.Config.add_defaults(
 	{
 		'smtp': {
-			"sender_email_address": "",
+			"from": "",
 			"host": "localhost",
 			"port": "",
 			"user": "",
@@ -33,7 +33,7 @@ class EmailOutputService(asab.Service, OutputABC):
 	def __init__(self, app, service_name="SmtpService", config_section_name='smtp'):
 		super().__init__(app, service_name)
 
-		self.Sender = asab.Config.get(config_section_name, "sender_email_address")
+		self.Sender = asab.Config.get(config_section_name, "from")
 		self.SSL = asab.Config.getboolean(config_section_name, "ssl")
 		self.StartTLS = asab.Config.getboolean(config_section_name, "starttls")
 		self.Host = asab.Config.get(config_section_name, "host")
@@ -82,12 +82,12 @@ class EmailOutputService(asab.Service, OutputABC):
 			assert (isinstance(email_cc, list))
 			msg['Cc'] = ', '.join(email_cc)
 
-		if email_subject is not None:
+		if email_subject is not None and len(email_from) > 0:
 			msg['Subject'] = email_subject
 		else:
 			msg['Subject'] = self.Subject
 
-		if email_from is not None:
+		if email_from is not None and len(email_from) > 0:
 			msg['From'] = sender = email_from
 		else:
 			msg['From'] = sender = self.Sender
