+++ etc/asab-iris.conf
@@ -5,11 +5,10 @@ providers=./library
 listen=:8080
 
 [smtp]
-sender_email_address=info@example.com
 host=smtp.example.com
 user=admin
 password=password
+from=info@example.com
 ssl=no
 starttls=yes
-subject='Alert!'
-file_size=500000000
+subject=Mail from ASAB Iris
